Add cancellation and timeout for calloc.

+++ scripts/wipe_data.sh
@@ -3,8 +3,8 @@
 echo "Usage: $0 mode(1:acct_table | 2:qos_table | 3:task_table | 4:user_table | 5:all | 6:acct_table+qos_table+user_table)"
 
 if [ "$#" -ne 1 ]; then
-    echo "Parameter error: please input mode num!"
-    exit 1
+  echo "Parameter error: please input mode num!"
+  exit 1
 fi
 
 mode=$1
@@ -23,9 +23,10 @@ port="27017"
 # 使用mongo shell连接到MongoDB服务器并清空指定的集合
 
 function wipe_collection() {
-    mongosh --username $username --password $password --host $host --port $port <<EOF
+  mongosh --username "$username" --password "$password" --host "$host" --port "$port" <<EOF
     use crane_db
     db.$1.deleteMany({})
+    exit
 EOF
 }
 
@@ -37,8 +38,11 @@ if [ "$mode" -eq 2 ] || [ "$mode" -eq 5 ] || [ "$mode" -eq 6 ]; then
 fi
 if [ "$mode" -eq 3 ] || [ "$mode" -eq 5 ]; then
   wipe_collection task_table
-  echo "remove file $unqpath"
-  rm "$unqpath"
+
+  if [ -e "$unqpath" ]; then
+    echo "Removing file $unqpath ..."
+    rm "$unqpath"
+  fi
 fi
 if [ "$mode" -eq 4 ] || [ "$mode" -eq 5 ] || [ "$mode" -eq 6 ]; then
   wipe_collection user_table
